Initial changes to Split E2E test
More change to split the test
