+++ tests/e2e/controlplane/control_plane_test.go
@@ -14,10 +14,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package operator
+package controlplane
 
 import (
-	"context"
 	"fmt"
 	"path/filepath"
 	"regexp"
@@ -27,55 +26,28 @@ import (
 	"github.com/istio-ecosystem/sail-operator/api/v1alpha1"
 	. "github.com/istio-ecosystem/sail-operator/pkg/test/util/ginkgo"
 	"github.com/istio-ecosystem/sail-operator/pkg/test/util/supportedversion"
-	. "github.com/istio-ecosystem/sail-operator/tests/e2e/operator/util/gomega"
-	"github.com/istio-ecosystem/sail-operator/tests/e2e/operator/util/helm"
-	"github.com/istio-ecosystem/sail-operator/tests/e2e/operator/util/kubectl"
+	common "github.com/istio-ecosystem/sail-operator/tests/e2e/util/common"
+	. "github.com/istio-ecosystem/sail-operator/tests/e2e/util/gomega"
+	"github.com/istio-ecosystem/sail-operator/tests/e2e/util/helm"
+	"github.com/istio-ecosystem/sail-operator/tests/e2e/util/kubectl"
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 	appsv1 "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
-	apiextensionsv1 "k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"sigs.k8s.io/controller-runtime/pkg/client"
 
 	"istio.io/istio/pkg/ptr"
 )
 
-var (
-	istiodVersionRegex = regexp.MustCompile(`Version:"(\d+\.\d+(\.\d+|-\w+))`)
-	sailCRDs           = []string{
-		// TODO: Find an alternative to this list
-		"authorizationpolicies.security.istio.io",
-		"destinationrules.networking.istio.io",
-		"envoyfilters.networking.istio.io",
-		"gateways.networking.istio.io",
-		"istiorevisions.operator.istio.io",
-		"istios.operator.istio.io",
-		"peerauthentications.security.istio.io",
-		"proxyconfigs.networking.istio.io",
-		"requestauthentications.security.istio.io",
-		"serviceentries.networking.istio.io",
-		"sidecars.networking.istio.io",
-		"telemetries.telemetry.istio.io",
-		"virtualservices.networking.istio.io",
-		"wasmplugins.extensions.istio.io",
-		"workloadentries.networking.istio.io",
-		"workloadgroups.networking.istio.io",
-	}
-)
+var istiodVersionRegex = regexp.MustCompile(`Version:"(\d+\.\d+(\.\d+|-\w+))`)
 
-var _ = Describe("Operator", Ordered, func() {
+var _ = Describe("Control Plane Installation", Ordered, func() {
 	SetDefaultEventuallyTimeout(120 * time.Second)
 	SetDefaultEventuallyPollingInterval(time.Second)
 
-	Describe("installation", func() {
-		// TODO: we  need to support testing both types of deployment for the operator, helm and olm via subscription.
-		// Discuss with the team if we should add a flag to the test to enable the olm deployment and don't do that deployment in different step
-		if skipDeploy {
-			Skip("Skipping the deployment of the operator")
-		}
+	Describe("given Istio version", func() {
 
-		BeforeAll(func() {
+		BeforeAll(func(ctx SpecContext) {
 			Expect(kubectl.CreateNamespace(namespace)).To(Succeed(), "Namespace failed to be created")
 
 			extraArg := ""
@@ -85,42 +57,12 @@ var _ = Describe("Operator", Ordered, func() {
 
 			Expect(helm.Install("sail-operator", filepath.Join(baseDir, "chart"), "--namespace "+namespace, "--set=image="+image, extraArg)).
 				To(Succeed(), "Operator failed to be deployed")
-		})
 
-		It("deploys all the CRDs", func(ctx SpecContext) {
-			Eventually(getList).WithArguments(ctx, cl, &apiextensionsv1.CustomResourceDefinitionList{}).Should(WithTransform(extractCRDNames, ContainElements(sailCRDs)),
-				"Not all Istio and Sail CRDs are present")
-			Success("Istio CRDs are present")
-		})
-
-		It("updates the CRDs status to Established", func(ctx SpecContext) {
-			for _, crdName := range sailCRDs {
-				Eventually(getObject).WithArguments(ctx, cl, key(crdName), &apiextensionsv1.CustomResourceDefinition{}).
-					Should(HaveCondition(apiextensionsv1.Established, metav1.ConditionTrue), "Error getting Istio CRD")
-			}
-			Success("CRDs are Established")
-		})
-
-		Specify("istio crd is present", func(ctx SpecContext) {
-			// When the operator runs in OCP cluster, the CRD is created but not available at the moment
-			Eventually(cl.Get).WithArguments(ctx, key("istios.operator.istio.io"), &apiextensionsv1.CustomResourceDefinition{}).
-				Should(Succeed(), "Error getting Istio CRD")
-			Success("Istio CRD is present")
-		})
-
-		It("starts successfully", func(ctx SpecContext) {
-			Eventually(getObject).WithArguments(ctx, cl, key(deploymentName, namespace), &appsv1.Deployment{}).
+			Eventually(common.GetObject).WithArguments(ctx, cl, common.Key(deploymentName, namespace), &appsv1.Deployment{}).
 				Should(HaveCondition(appsv1.DeploymentAvailable, metav1.ConditionTrue), "Error getting Istio CRD")
+			Success("Operator is deployed in the namespace and Running")
 		})
 
-		AfterAll(func() {
-			if CurrentSpecReport().Failed() {
-				LogDebugInfo()
-			}
-		})
-	})
-
-	Describe("given Istio version", func() {
 		for _, version := range supportedversion.List {
 			// Note: This var version is needed to avoid the closure of the loop
 			version := version
@@ -142,7 +84,7 @@ spec:
   version: %s
   namespace: %s`
 						yaml = fmt.Sprintf(yaml, version.Name, istioCniNamespace)
-						log("IstioCNI YAML:", indent(2, yaml))
+						Log("IstioCNI YAML:", indent(2, yaml))
 						Expect(kubectl.ApplyString(yaml)).To(Succeed(), "IstioCNI creation failed")
 						Success("IstioCNI created")
 					})
@@ -150,7 +92,7 @@ spec:
 					It("deploys the CNI DaemonSet", func(ctx SpecContext) {
 						Eventually(func(g Gomega) {
 							daemonset := &appsv1.DaemonSet{}
-							g.Expect(cl.Get(ctx, key("istio-cni-node", istioCniNamespace), daemonset)).To(Succeed(), "Error getting IstioCNI DaemonSet")
+							g.Expect(cl.Get(ctx, common.Key("istio-cni-node", istioCniNamespace), daemonset)).To(Succeed(), "Error getting IstioCNI DaemonSet")
 							g.Expect(daemonset.Status.NumberAvailable).
 								To(Equal(daemonset.Status.CurrentNumberScheduled), "CNI DaemonSet Pods not Available; expected numberAvailable to be equal to currentNumberScheduled")
 						}).Should(Succeed(), "CNI DaemonSet Pods are not Available")
@@ -158,13 +100,13 @@ spec:
 					})
 
 					It("updates the status to Reconciled", func(ctx SpecContext) {
-						Eventually(getObject).WithArguments(ctx, cl, key(istioCniName), &v1alpha1.IstioCNI{}).
+						Eventually(common.GetObject).WithArguments(ctx, cl, common.Key(istioCniName), &v1alpha1.IstioCNI{}).
 							Should(HaveCondition(v1alpha1.IstioCNIConditionReconciled, metav1.ConditionTrue), "IstioCNI is not Reconciled; unexpected Condition")
 						Success("IstioCNI is Reconciled")
 					})
 
 					It("updates the status to Ready", func(ctx SpecContext) {
-						Eventually(getObject).WithArguments(ctx, cl, key(istioCniName), &v1alpha1.IstioCNI{}).
+						Eventually(common.GetObject).WithArguments(ctx, cl, common.Key(istioCniName), &v1alpha1.IstioCNI{}).
 							Should(HaveCondition(v1alpha1.IstioCNIConditionReady, metav1.ConditionTrue), "IstioCNI is not Ready; unexpected Condition")
 						Success("IstioCNI is Ready")
 					})
@@ -187,26 +129,26 @@ spec:
   version: %s
   namespace: %s`
 						istioYAML = fmt.Sprintf(istioYAML, version.Name, controlPlaneNamespace)
-						log("Istio YAML:", indent(2, istioYAML))
+						Log("Istio YAML:", indent(2, istioYAML))
 						Expect(kubectl.ApplyString(istioYAML)).
 							To(Succeed(), "Istio CR failed to be created")
 						Success("Istio CR created")
 					})
 
 					It("updates the Istio CR status to Reconciled", func(ctx SpecContext) {
-						Eventually(getObject).WithArguments(ctx, cl, key(istioName), &v1alpha1.Istio{}).
+						Eventually(common.GetObject).WithArguments(ctx, cl, common.Key(istioName), &v1alpha1.Istio{}).
 							Should(HaveCondition(v1alpha1.IstioConditionReconciled, metav1.ConditionTrue), "Istio is not Reconciled; unexpected Condition")
 						Success("Istio CR is Reconciled")
 					})
 
 					It("updates the Istio CR status to Ready", func(ctx SpecContext) {
-						Eventually(getObject).WithArguments(ctx, cl, key(istioName), &v1alpha1.Istio{}).
+						Eventually(common.GetObject).WithArguments(ctx, cl, common.Key(istioName), &v1alpha1.Istio{}).
 							Should(HaveCondition(v1alpha1.IstioConditionReady, metav1.ConditionTrue), "Istio is not Ready; unexpected Condition")
 						Success("Istio CR is Ready")
 					})
 
 					It("deploys istiod", func(ctx SpecContext) {
-						Eventually(getObject).WithArguments(ctx, cl, key("istiod", controlPlaneNamespace), &appsv1.Deployment{}).
+						Eventually(common.GetObject).WithArguments(ctx, cl, common.Key("istiod", controlPlaneNamespace), &appsv1.Deployment{}).
 							Should(HaveCondition(appsv1.DeploymentAvailable, metav1.ConditionTrue), "Istiod is not Available; unexpected Condition")
 						Expect(getVersionFromIstiod()).To(Equal(version.Version), "Unexpected istiod version")
 						Success("Istiod is deployed in the namespace and Running")
@@ -226,9 +168,9 @@ spec:
 					})
 
 					It("removes everything from the namespace", func(ctx SpecContext) {
-						Eventually(cl.Get).WithArguments(ctx, key("istiod", controlPlaneNamespace), &appsv1.Deployment{}).
+						Eventually(cl.Get).WithArguments(ctx, common.Key("istiod", controlPlaneNamespace), &appsv1.Deployment{}).
 							Should(ReturnNotFoundError(), "Istiod should not exist anymore")
-						checkNamespaceEmpty(ctx, controlPlaneNamespace)
+						common.CheckNamespaceEmpty(ctx, cl, controlPlaneNamespace)
 						Success("Namespace is empty")
 					})
 				})
@@ -241,9 +183,9 @@ spec:
 
 					It("removes everything from the CNI namespace", func(ctx SpecContext) {
 						daemonset := &appsv1.DaemonSet{}
-						Eventually(cl.Get).WithArguments(ctx, key("istio-cni-node", istioCniNamespace), daemonset).
+						Eventually(cl.Get).WithArguments(ctx, common.Key("istio-cni-node", istioCniNamespace), daemonset).
 							Should(ReturnNotFoundError(), "IstioCNI DaemonSet should not exist anymore")
-						checkNamespaceEmpty(ctx, istioCniNamespace)
+						common.CheckNamespaceEmpty(ctx, cl, istioCniNamespace)
 						Success("CNI namespace is empty")
 					})
 				})
@@ -252,7 +194,7 @@ spec:
 
 		AfterAll(func(ctx SpecContext) {
 			if CurrentSpecReport().Failed() {
-				LogDebugInfo()
+				common.LogDebugInfo()
 			}
 
 			By("Cleaning up the Istio namespace")
@@ -261,23 +203,21 @@ spec:
 			By("Cleaning up the IstioCNI namespace")
 			Expect(cl.Delete(ctx, &corev1.Namespace{ObjectMeta: metav1.ObjectMeta{Name: istioCniNamespace}})).To(Succeed(), "IstioCNI Namespace failed to be deleted")
 
+			By("Deleting any left-over Istio and IstioRevision resources")
+			Expect(forceDeleteIstioResources()).To(Succeed())
+			Success("Resources deleted")
 			Success("Cleanup done")
 		})
 	})
 
 	AfterAll(func() {
-		By("Deleting any left-over Istio and IstioRevision resources")
-		Expect(forceDeleteIstioResources()).To(Succeed())
-		Success("Resources deleted")
-
-		By("Uninstalling the operator")
+		By("Deleting operator deployment")
 		Expect(helm.Uninstall("sail-operator", "--namespace "+namespace)).
 			To(Succeed(), "Operator failed to be deleted")
-		Success("Operator uninstalled")
+		GinkgoWriter.Println("Operator uninstalled")
 
-		By("Deleting the CRDs")
-		Expect(kubectl.DeleteCRDs(sailCRDs)).To(Succeed(), "CRDs failed to be deleted")
-		Success("CRDs deleted")
+		Expect(kubectl.DeleteNamespace(namespace)).To(Succeed(), "Namespace failed to be deleted")
+		Success("Namespace deleted")
 	})
 })
 
@@ -292,6 +232,11 @@ func getVersionFromIstiod() (string, error) {
 		return matches[1], nil
 	}
 	return "", fmt.Errorf("error getting version from istiod: version not found in output: %s", output)
+}
+
+func indent(level int, str string) string {
+	indent := strings.Repeat(" ", level)
+	return indent + strings.ReplaceAll(str, "\n", "\n"+indent)
 }
 
 func forceDeleteIstioResources() error {
@@ -313,99 +258,4 @@ func forceDeleteIstioResources() error {
 	}
 
 	return nil
-}
-
-func indent(level int, str string) string {
-	indent := strings.Repeat(" ", level)
-	return indent + strings.ReplaceAll(str, "\n", "\n"+indent)
-}
-
-func log(a ...any) {
-	GinkgoWriter.Println(a...)
-}
-
-func extractCRDNames(crdList *apiextensionsv1.CustomResourceDefinitionList) []string {
-	var names []string
-	for _, crd := range crdList.Items {
-		names = append(names, crd.ObjectMeta.Name)
-	}
-	return names
-}
-
-func LogDebugInfo() {
-	// General debugging information to help diagnose the failure
-	// TODO: Add more debugging information for others resources
-	resource, err := kubectl.GetYAML(controlPlaneNamespace, "istio", istioName)
-	if err != nil {
-		GinkgoWriter.Println("Error getting Istio CR: ", err)
-	}
-	GinkgoWriter.Println("Istio CR: \n", resource)
-
-	output, err := kubectl.GetPods(controlPlaneNamespace, "-o wide")
-	if err != nil {
-		GinkgoWriter.Println("Error getting pods: ", err)
-	}
-	GinkgoWriter.Println("Pods in Istio CR namespace: \n", output)
-
-	logs, err := kubectl.Logs(namespace, "deploy/"+deploymentName, ptr.Of(120*time.Second))
-	if err != nil {
-		GinkgoWriter.Println("Error getting logs from the operator: ", err)
-	}
-	GinkgoWriter.Println("Logs from sail-operator pod: \n", logs)
-
-	logs, err = kubectl.Logs(controlPlaneNamespace, "deploy/istiod", ptr.Of(120*time.Second))
-	if err != nil {
-		GinkgoWriter.Println("Error getting logs from the istiod: ", err)
-	}
-	GinkgoWriter.Println("Logs from istiod pod: \n", logs)
-}
-
-// key returns the client.ObjectKey for the given name and namespace. If no namespace is provided, it returns a key cluster scoped
-func key(name string, namespace ...string) client.ObjectKey {
-	if len(namespace) > 1 {
-		panic("you can only provide one namespace")
-	} else if len(namespace) == 1 {
-		return client.ObjectKey{Name: name, Namespace: namespace[0]}
-	}
-	return client.ObjectKey{Name: name}
-}
-
-// getObject returns the object with the given key
-func getObject(ctx context.Context, cl client.Client, key client.ObjectKey, obj client.Object) (client.Object, error) {
-	err := cl.Get(ctx, key, obj)
-	return obj, err
-}
-
-// getList invokes client.List and returns the list
-func getList(ctx context.Context, cl client.Client, list client.ObjectList, opts ...client.ListOption) (client.ObjectList, error) {
-	err := cl.List(ctx, list, opts...)
-	return list, err
-}
-
-// checkNamespaceEmpty checks if the given namespace is empty
-func checkNamespaceEmpty(ctx SpecContext, ns string) {
-	// TODO: Check to add more validations
-	Eventually(func() ([]corev1.Pod, error) {
-		podList := &corev1.PodList{}
-		err := cl.List(ctx, podList, client.InNamespace(ns))
-		return podList.Items, err
-	}).Should(BeEmpty(), "No pods should be present in the namespace")
-
-	Eventually(func() ([]appsv1.Deployment, error) {
-		deploymentList := &appsv1.DeploymentList{}
-		err := cl.List(ctx, deploymentList, client.InNamespace(ns))
-		return deploymentList.Items, err
-	}).Should(BeEmpty(), "No Deployments should be present in the namespace")
-
-	Eventually(func() ([]appsv1.DaemonSet, error) {
-		daemonsetList := &appsv1.DaemonSetList{}
-		err := cl.List(ctx, daemonsetList, client.InNamespace(ns))
-		return daemonsetList.Items, err
-	}).Should(BeEmpty(), "No DaemonSets should be present in the namespace")
-
-	Eventually(func() ([]corev1.Service, error) {
-		serviceList := &corev1.ServiceList{}
-		err := cl.List(ctx, serviceList, client.InNamespace(ns))
-		return serviceList.Items, err
-	}).Should(BeEmpty(), "No Services should be present in the namespace")
 }
