+++ tests/e2e/util/kubectl/resourcecondition_types.go
