+++ tests/e2e/util/kubectl/resourcelist_types.go
